+++ src/chains/Cosmoshub/index.ts
@@ -1,4 +1,5 @@
-import { AssetInfo, Chain, ChainInfo } from "../../interface";
+import { AssetInfo } from "../../assets/types";
+import { Chain, ChainInfo } from "../types";
 import { bech32 } from "bech32";
 import Axelar from "../Axelar";
