+++ src/chains/types/index.ts
@@ -1,9 +1,6 @@
-import { SourceOrDestination } from "./Miscellaneous";
-import { SocketServices } from "../services/SocketServices";
-import { AssetInfoResponse } from "./AssetTransferObject";
-
-export * from "./AssetTransferObject";
-export * from "./Miscellaneous";
+import { AssetInfo } from "../../assets/types";
+import { SocketServices } from "../../services";
+import { SourceOrDestination } from "../../services/types";
 
 export interface Chain {
   chainInfo: ChainInfo;
@@ -57,17 +54,6 @@ export interface ChainInfo {
   };
 }
 
-export interface AssetInfo {
-  assetSymbol?: string;
-  assetName?: string;
-  assetAddress?: string;
-  common_key?: string;
-  fullySupported?: boolean;
-  native_chain?: string;
-  minDepositAmt?: number;
-  decimals?: number;
-}
-
 export type BlockchainWaitingServiceFinder = (
   chainInfo: ChainInfo,
   assetInfo: AssetInfo,
@@ -75,3 +61,56 @@ export type BlockchainWaitingServiceFinder = (
   environment: string,
   roomId?: string
 ) => BlockchainWaitingService | Promise<BlockchainWaitingService>;
+
+export interface AssetTransferObject {
+  sourceChainInfo: ChainInfo;
+  destinationChainInfo: ChainInfo;
+  selectedSourceAsset: AssetInfo;
+  selectedDestinationAsset: AssetInfo;
+  signature: string;
+  otc: string;
+  publicAddr: string;
+  transactionTraceId?: string;
+}
+
+export interface AssetInfoWithTrace extends AssetInfo {
+  traceId: string;
+  assetInfo: AssetInfo;
+}
+
+export interface AssetInfoResponse extends AssetInfo {
+  sourceOrDestChain: SourceOrDestination;
+  traceId: string;
+}
+
+export enum LinkType {
+  EVM = "/evm.v1beta1.LinkRequest",
+  COS = "/axelarnet.v1beta1.LinkRequest",
+}
+
+export interface LinkRequestBody {
+  "@type": LinkType;
+  recipient_addr: string;
+  recipient_chain: string;
+}
+
+export interface EVMLinkRequestBody extends LinkRequestBody {
+  chain: string; //source chain
+  asset: string;
+}
+
+export interface COSLinkRequestBody extends LinkRequestBody {
+  asset: string;
+}
+
+// for connections from ui >> bridge server
+export enum SocketListenerTypes {
+  /*axelarnet listener for deposit event*/
+  LINK = "LINK",
+  WAIT_FOR_DEPOSIT = "WAIT_FOR_DEPOSIT",
+  DEPOSIT_CONFIRMED = "DEPOSIT_CONFIRMED",
+}
+
+export interface SocketListenerTopic {
+  topic: SocketListenerTypes;
+}
