+++ src/index.ts
@@ -1,6 +1,9 @@
 export * from "./assets";
 export * from "./chains";
 export * from "./constants";
-export * from "./interface";
 export * from "./libs";
 export * from "./utils";
+
+export * from "./assets/types";
+export * from "./chains/types";
+export * from "./services/types";
