--- src/types/otc.type.ts
@@ -1,4 +1,0 @@
-export interface OTC {
-  otc: string;
-  validationMsg: string;
-}
